feat: move path resolutions to package file
