+++ apps/frontend/package.json
@@ -10,7 +10,26 @@
     "lint": "next lint && prettier --log-level=warn ./src --check",
     "format": "prettier --log-level=warn ./src --write",
     "dev": "next dev --port 8001",
-    "codegen": "npm run build-codegen && node ./scripts/codegen && prettier --log-level=warn ./codegen/output --write"
+    "codegen": "npm run build-codegen && node ./scripts/codegen.mjs && prettier --log-level=warn ./codegen/output --write"
+  },
+  "imports": {
+    "#environment": "./src/environment/index.ts",
+    "#environment/*": "./src/environment/*/index.ts",
+    "#pages/types": "./src/app/types.ts",
+    "#lib/*": "./src/lib/*/index.ts",
+    "#server": "./src/server/index.ts",
+    "#server/*": "./src/server/*/index.ts",
+    "#store/*": "./src/browser/store/*/index.ts",
+    "#browser/*": "./src/browser/*",
+    "#hooks": "./src/hooks/index.ts",
+    "#components": "./src/components/index.ts",
+    "#components/types": "./src/components/types.ts",
+    "#components/*": "./src/components/*/index.ts",
+    "#entities/*": "./src/entities/*/index.ts",
+    "#types": "./src/types.ts",
+    "#json-schema/*": [
+      "./codegen/output/json-schema/*/index.ts"
+    ]
   },
   "dependencies": {
     "@hyperjump/json-schema": "^1.9.9",
